requirements.txt
asgiref==3.10.0
azure-core==1.35.1
azure-identity==1.25.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.2
Django==5.2.7
gunicorn==23.0.0
idna==3.11
msal==1.34.0
msal-extensions==1.3.1
packaging==25.0
psycopg2-binary==2.9.11
pycparser==2.23
PyJWT==2.10.1
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.15.0
urllib3==2.5.0
whitenoise==6.11.0
